added a services model to make it more readable
